emol/requirements.txt
Authomatic==0.1.0.post1
Jinja2==2.10
Werkzeug==0.14.1
alembic==0.9.9
Flask_RESTful==0.3.5
Flask_Login==0.4.0
Flask_SQLAlchemy==2.2
pytest==3.3.1
Flask_Migrate==2.1.1
pytz==2016.7
setuptools==20.7.0
Flask_Script==2.0.6
Flask==1.0.1
SQLAlchemy==1.2.7
pycrypto==2.6.1
python_dateutil==2.7.2
unicode_slugify==0.1.3
